refactor(server): break up massive MapManager.ts into logical modules
- Split 4,497-line MapManager.ts into 8 focused modules:
  - MapTypes.ts: Core types, enums, and interfaces
  - FractalHeightGenerator.ts: Height map generation algorithms
  - TemperatureMap.ts: Climate and temperature mapping
  - TerrainSelectionEngine.ts: Climate-aware terrain selection
  - IslandGenerator.ts: Island-based terrain generation
  - RiverGenerator.ts: Advanced river placement system
  - ResourceGenerator.ts: Resource distribution logic
  - StartingPositionGenerator.ts: Player spawn evaluation

- Reduced MapManager from 4,497 to 532 lines (89% reduction)
- Maintains all existing functionality and APIs
- Improves maintainability, testability, and readability
- Each module has single responsibility principle
- All TypeScript compilation and linting passes
